requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
einops==0.4.1
exceptiongroup==1.1.0
filelock==3.9.0
fonttools==4.38.0
huggingface-hub==0.12.0
idna==3.4
iniconfig==2.0.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.9.3
torchvision==0.12.0
tqdm==4.64.1
transformers==4.21.3
typing_extensions==4.4.0
urllib3==1.26.14
warmup-scheduler==0.3.2
